+++ packages/components/src/ProForm/hooks/useForm.tsx
@@ -1,22 +1,23 @@
 import { useRef } from 'react';
 
 const useForm = () => {
-  const clickRef = useRef<any>();
   const formRef = useRef<any>();
 
   // 触发验证表单
-  const submitvalidate = () => clickRef.current.click();
+  const submitvalidate = () => formRef.current.submitvalidate();
   // 重置表单
   const resetForm = () => formRef.current.resetForm();
   // 获取表单值
-  const getFormValues = () => formRef.current.current;
+  const getFieldValues = () => formRef.current.getFieldValues();
+  // 获取错误信息
+  const getErrors = () => formRef.current?.errors || {};
 
   return {
     formRef,
-    clickRef,
     submitvalidate,
     resetForm,
-    getFormValues,
+    getFieldValues,
+    getErrors,
   };
 };